In sync

+++ lib/Effects/Effects/FadeSparkle.h
@@ -8,7 +8,7 @@ void fadeSparkleEffectLoop(struct EffectDataPacket *effectState, bool isFadeDown
 
   const uint8_t sparkleRefreshHz = LED_CNT * 5;
   const uint8_t fadeRefreshHz = sparkleRefreshHz;
-  const uint16_t fadeSparkleLoopMs = 600000;
+  const uint32_t fadeSparkleLoopMs = 600000;
   const uint8_t loopPosition = effectState->loopPosition % fadeSparkleLoopMs;
 
   const uint8_t hue1 = 128;
